SentimentEscalator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
requests
scikit-learn
numpy
transformers
torch
accelerate
sentencepiece
pandas
datasets
evaluate
safetensors